# src/main/resources/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: NotifyMe API
  version: 1.0.0
security:
  - oAuth2ClientCredentials: []

paths:
  /v1/login:
    post:
      summary: API de autenticação
      description: Para utilizar as APIs é necessário logar e pegar um token.
      tags:
        - Autenticacao
      operationId: postLoginV1
      requestBody:
        description: >
          Informe os dados - Fields **username** e **password** para pegar um
          token da aplicação.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResposeDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/usuario:
    post:
      summary: API de inclusão de novo usuário
      description: Para utilizar as APIs é necessário logar e pegar um token.
      tags:
        - Usuario
      operationId: postUsuarioV1
      requestBody:
        description: Cadastro de novo usuario.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequestDTO'
      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: API de consultar os usuarios
      description: Essa API busca todos os usuarios do condominio.
      tags:
        - Usuario
      operationId: getCondominioV1
      parameters:
        - $ref: '#/components/parameters/authorization'
        - $ref: '#/components/parameters/requiredPageSize'
        - $ref: '#/components/parameters/requiredPageNumber'
        - $ref: '#/components/parameters/nomeInQuery'
        - $ref: '#/components/parameters/cpfInQuery'
        - $ref: '#/components/parameters/telefoneInQuery'
        - $ref: '#/components/parameters/emailInQuery'
        - $ref: '#/components/parameters/statusInQuery'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuariosResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/usuario/valida-token:
    post:
      summary: Valida um token de ativação e ativa o usuário
      description: Verifica se o token é válido e ativa o cadastro do usuário.
      tags:
        - Usuario
      operationId: postUsuarioValidaTokenV1
      requestBody:
        description: Informe o token recebido via e-mail para validar seu cadastro.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: integer
                  description: Token único para validação.
                  example: 101020
              required:
                - token
      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/usuario/redefinir-senha:
    post:
      summary: Redefinir a senha do usuário
      description: Será gerado uma nova senha para redefinição.
      tags:
        - Usuario
      operationId: postUsuarioRedefinirSenhaV1
      requestBody:
        description: Informe o e-mail ou o telefone para receber o e-mail com o token para a troca da senha.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contato:
                  type: string
                  description: E-mail ou telefone da sua conta.
                  example: teste@teste.com
              required:
                - contato
      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/usuario/{id}/nova-senha:
    put:
      summary: Criar uma nova senha
      description: Criar uma nova senha para o usuário.
      tags:
        - Usuario
      operationId: postUsuarioNovaSenhaV1
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser atualizado.
          schema:
            type: integer
            example: "12355"
      requestBody:
        description: Informe os valores para criar uma nova senha para o usuário.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaSenhaRequestDTO'

      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/usuario/{id}:
    put:
      summary: API de edição de usuário
      description: Para utilizar as APIs é necessário logar e pegar um token.
      tags:
        - Usuario
      operationId: putUsuarioV1
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser atualizado.
          schema:
            type: integer
            example: 101020
      requestBody:
        description: Edição de usuário.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUsuarioRequestDTO'

      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/usuario/deletar/{id}:
    delete:
      summary: API de edição de usuário
      description: Para utilizar as APIs é necessário logar e pegar um token.
      tags:
        - Usuario
      operationId: deleteUsuario
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser atualizado.
          schema:
            type: integer
            example: 102020

      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/condominio:
    put:
      summary: Editar o condominio condominio
      description: Essa API cadastra condominio.
      tags:
        - Condominio
      operationId: putCondominioV1
      parameters:
        - $ref: '#/components/parameters/authorization'
      requestBody:
        description: Criar novo condominio.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CondominioBaseRequestDTO'
      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: API consulta de condominio
      description: Essa API busca todos os condomínios.
      tags:
        - Condominio
      operationId: getUsuariosV1
      parameters:
        - $ref: '#/components/parameters/authorization'
        - $ref: '#/components/parameters/requiredPageSize'
        - $ref: '#/components/parameters/requiredPageNumber'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CondominioResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'


  /v1/condomino:
    post:
      summary: Cadastro do condominos
      description: Cadastrar todos os condominos do condominio.
      tags:
        - Condomino
      operationId: postCondominoV1
      requestBody:
        description: Criar novo condomino.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CondominoRequestDTO'
      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/cep:
    get:
      summary: API de autenticação
      description: Para utilizar as APIs é necessário logar e pegar um token.
      tags:
        - cep
      operationId: getCepV1
      parameters:
        - $ref: '#/components/parameters/estado'
        - $ref: '#/components/parameters/cidade'
        - $ref: '#/components/parameters/endereco'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CepResponseDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/upload/foto-perfil/{id}:
    post:
      summary: API de edição de usuário
      description: Para utilizar as APIs é necessário logar e pegar um token.
      tags:
        - Upload
      operationId: upFotoPerfilV1
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário a ser atualizado.
          schema:
            type: string
            example: 102020

      requestBody:
        description: Edição de usuário.
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FotoPerfilRequestDTO'

      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/encomenda/upload:
    post:
      summary: API inicial de chegada de encomenda
      description: API que identifica o destinatario e realiza o envio de notificação.
      tags:
        - Encomenda
      operationId: upImagemEncomenda
      requestBody:
        description: Envio de imagem para identificar o destinatario da encomenda.
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImagemEncomendaRequestDTO'

      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/unidade:
    post:
      summary: Cadastrar as unidades
      description: Casdastra as unidades do condominio
      tags:
        - Unidade
      operationId: postUnidadeV1
      requestBody:
        description: Criar novo condominio.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnidadeRequestDTO'
      responses:
        '200':
          description: Success response
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      description: >-
        Request an access token to Keycloak. Then use the obtained token to
        access secured endpoints.
      flows:
        clientCredentials:
          tokenUrl: >-
            https://your-keycloak-domain/realms/{realm}/protocol/openid-connect/token
          scopes: {}
  schemas:
    LoginRequestDTO:
      type: object
      properties:
        username:
          description: Nome de usuário ou telefone.
          type: string
          example: '18998224242'
        password:
          description: Informe a senha cadastrada.
          type: string
          example: '@df4fr4rf4r'
      required:
        - username
        - password

    LoginResposeDTO:
      type: object
      properties:
        accessToken:
          description: Token gerado para utilizar nas APIs.
          type: string
          example: >
            eyJhbGciOiJSUzI1NieyJpc3MiOiJ2aWxhZ2JhY2tlbmQiLCJzdWIiOiIxIiwiZXhwIjoxN
        expiresIn:
          description: Tempo até expiração.
          type: integer
          example: 10800
      required:
        - accessToken
        - expiresIn

    UsuarioRequestDTO:
      type: object
      properties:
        nome:
          description: Nome de usuário.
          type: string
          example: 'Jose da Silva Santos'
        telefone:
          description: Informe um telefone com DDD.
          type: string
          example: '189969692020'
        email:
          description: Informe um e-mail válido.
          type: string
          format: email
          example: 'notifyme@teste.com'
        password:
          description: Informe uma senha.
          type: string
          example: '@FacaS3mP0ntA'
        cpf:
          description: Informe um CPF válido.
          pattern: "^[0-9]{11}$"
          type: string
          example: '79222300017'
          minLength: 11
          maxLength: 11
        condominio:
          $ref: '#/components/schemas/CondominioSimplesDTO'
      required:
        - nome
        - email
        - telefone
        - password
        - cpf
        - condominio

    UpdateUsuarioRequestDTO:
      type: object
      properties:
        nome:
          description: Nome de usuário.
          type: string
          example: 'Jose da Silva Santos'
        telefone:
          description: Informe um telefone com DDD.
          type: string
          example: '189969692020'
        email:
          description: Informe um e-mail valido.
          type: string
          example: 'notifyme@teste.com'
        cpf:
          description: Informe o cpf valido.
          type: string
          example: '999.999.999-99'
        password:
          description: Informe um novo password.
          type: string
          example: '********'
        role:
          description: Informe uma Role.
          type: string
          example: 'ADMIN'
        status:
          description: Informe uma Role.
          type: integer
          example: '1'
      required:
        - id

    FotoPerfilRequestDTO:
      type: object
      properties:
        fotoPerfil:
          description: Foto de Perfil de usuário.
          type: string
          format: binary
          example: 'https://editor.swagger.io/'
      required:
        - fotoPerfil
    ImagemEncomendaRequestDTO:
      type: object
      properties:
        imagemEncomenda:
          description: Foto da encomenda.
          type: string
          format: binary
          example: 'https://editor.swagger.io/'
      required:
        - imagemEncomenda

    CepResponseDTO:
      type: object
      properties:
        cep:
          description: Código postal.
          type: string
          example: '01001000'
        logradouro:
          description: Logradouro.
          type: string
          example: 'Praça da Sé'
        complemento:
          description: Complemento.
          type: string
          example: 'lado ímpar'
        bairro:
          description: Bairro.
          type: string
          example: 'Sé'
        localidade:
          description: Cidade.
          type: string
          example: 'São Paulo'
        uf:
          description: Unidade Federativa.
          type: string
          example: 'SP'
    NovaSenhaRequestDTO:
      type: object
      properties:
        senha:
          description: Informe a nova senha.
          type: string
          example: '#1TestSenh@'
        token:
          description: Token que recebeu via e-mail.
          type: integer
          example: 101020
      required:
        - senha
        - token

    CondominioBaseRequestDTO:
      type: object
      properties:
        razao:
          description: Nome do condominio.
          type: string
          example: 'Residencial Solar das Palmeiras'
        cnpj:
          description: CNPJ do condominio.
          type: string
          pattern: "^[0-9]{14}$"
          minLength: 14
          maxLength: 14
          example: '44473179000170'
        telefone:
          description: Telefone do condominio.
          type: string
          example: '18998232323'
        email:
          description: E-mail do condominio.
          type: string
          format: email
          example: 'teste@teste.com'
        cep:
          description: CEP do condominio.
          type: string
          example: '19800-000'
        logradouro:
          description: Logradouro do condominio.
          type: string
          example: 'Rua marechal Rondon 101'
        numero:
          description: Logradouro do condominio.
          type: string
          example: '101'
        complemento:
          description: Complemento.
          type: string
          example: 'Perto da praça da Paz'
        cidade:
          description: Cidade do condominio.
          type: string
          example: 'Maringa'
        estado:
          description: Estado do condominio.
          type: string
          example: 'SP'
        bairro:
          description: Bairro do condominio.
          type: string
          example: 'Vila Brasileira'
        tipo:
          description: Tipo do condomínio.
          type: string
          enum:
            - RESIDENCIAL
            - COMERCIAL
          example: 'RESIDENCIAL'
      required:
        - nome
        - cnpj
        - telefone
        - email
        - cep
        - logradouro
        - numero
        - cidade
        - estado
        - bairro
        - tipo

    CondominioSimplesDTO:
      type: object
      properties:
        razao:
          description: Nome do condomínio.
          type: string
          example: 'Residencial Solar das Palmeiras'
        cnpj:
          description: CNPJ do condomínio.
          type: string
          pattern: "^[0-9]{14}$"
          minLength: 14
          maxLength: 14
          example: '44473179000170'
      required:
        - nome
        - cnpj

    UnidadeRequestDTO:
      type: object
      properties:
        condominioId:
          description: ID do condominio
          type: integer
          example: 101020
        descricao:
          description: descricao.
          type: string
          example: 'Descricao da unidade'

      required:
        - condominioId
        - descricao
    CondominoRequestDTO:
      type: object
      properties:
        unidadeId:
          description: Id da Unidade
          type: integer
          example: 102020
        nome:
          description: Nome do condominio.
          type: string
          example: 'Residencial Solar das Palmeiras'
        telefone:
          description: CNPJ do condominio.
          type: string
          example: '44473179000170'
        email:
          description: Telefone do condominio.
          type: string
          format: email
          example: '18998232323'

      required:
        - unidadeId
        - nome
        - telefone
        - email

    PageResult:
      type: object
      properties:
        pageable:
          $ref: '#/components/schemas/PageableResult'
        totalPages:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 1
        pageNumber:
          type: integer
          example: 1
        numberOfElements:
          type: integer
          example: 1
        totalElements:
          type: integer
          format: int64
          example: 1
        last:
          type: boolean
          example: true
        first:
          type: boolean
          example: true
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: true
        empty:
          type: boolean
          example: true
        sort:
          $ref: '#/components/schemas/SortResult'

    PageableResult:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 5
        offset:
          type: integer
          format: int64
          example: 0
        unpaged:
          type: boolean
          example: false
        paged:
          type: boolean
          example: true

    SortResult:
      type: object
      properties:
        sorted:
          type: boolean
          example: true
        unsorted:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false

    Errors:
      type: object
      description: A collection of errors when processing a request.
      properties:
        error:
          type: array
          items:
            $ref: '#/components/schemas/Error'

    Error:
      type: object
      description: An error when processing a request.
      properties:
        code:
          description: The error code.
          type: string
          example: NotNull
        message:
          description: The error message.
          type: string
          example: 'receiver.receiverTaxId: cannot be null'
        field:
          description: The field or attribute the error is related to.
          type: string
          example: receiver.receiverTaxId
        action:
          description: 'Action to solve the issue.'
          type: string
      required:
        - code
        - message

    UsuariosResponseWrapper:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UsuariosInfoListResponse'
    UsuariosInfoListResponse:
      allOf:
        - type: object
          properties:
            usuarios:
              type: array
              items:
                $ref: '#/components/schemas/UsuariosResponse'
        - $ref: '#/components/schemas/PageResult'

    CondominioResponseWrapper:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CondominioInfoListResponse'

    CondominioInfoListResponse:
      allOf:
        - type: object
          properties:
            condominios:
              type: array
              items:
                $ref: '#/components/schemas/CondominioResponse'
        - $ref: '#/components/schemas/PageResult'
    UsuariosResponse:
      type: object
      properties:
        id:
          type: integer
          description: Id do Condominio.
          example: '12345678910'
        nome:
          type: string
          description: Nome do Condominio
          example: 'Condominio Solar das Palmeiras'
        foto:
          type: string
          description: Foto do usuário
          example: 'Foto do usuário'
        cpf:
          type: string
          description: CNPJ do Solar das Palmeiras.
          example: '31577199000149'
        telefone:
          type: string
          description: Telefone do Codominio.
          example: '18998237470'
        email:
          type: string
          description: Telefone do Codominio.
          example: '18998237470'
        status:
          $ref: '#/components/schemas/StatusUsuario'

    CondominioResponse:
      type: object
      properties:
        condominioId:
          type: integer
          description: Id do Condominio.
          example: 102020
        condominioNome:
          type: string
          description: Nome do Condominio
          example: 'Condominio Solar das Palmeiras'
        condominioCnpj:
          type: string
          description: CNPJ do Solar das Palmeiras.
          example: '31577199000149'
        condominioTelefone:
          type: string
          description: Telefone do Codominio.
          example: '18998237470'
        condominioEmail:
          type: string
          description: Telefone do Codominio.
          example: '18998237470'
        condominioCep:
          type: string
          description: Telefone do Codominio.
          example: '18998237470'
        condominioLogradouro:
          type: string
          description: Telefone do Codominio.
          example: '18998237470'

    StatusUsuario:
      description: Indicates the status of monitoring
      type: string
      enum: [ATIVO, INATIVO, BLOQUEADO, PENDENTE_DE_VALIDACAO]
      example: ATIVO

  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'

    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'

  parameters:
    requiredPageSize:
      name: pageSize
      in: query
      description: Size of the pages as required.
      schema:
        type: integer
        minimum: 1
      required: true
      example: 25

    requiredPageNumber:
      name: pageNumber
      in: query
      description: Number of the current page as required.
      schema:
        type: integer
        minimum: 0
      required: true
      example: 3

    estado:
      name: estado
      in: query
      description: Estado.
      schema:
        type: string
      required: true

    cidade:
      name: cidade
      in: query
      description: Cidade.
      schema:
        type: string
      required: true

    endereco:
      name: endereco
      in: query
      description: Endereço.
      schema:
        type: string
      required: true

    nomeInQuery:
      name: nome
      in: query
      description: Nome.
      schema:
        type: string
      required: false
    cpfInQuery:
      name: cpf
      in: query
      description: cpf.
      schema:
        type: string
      required: false
    telefoneInQuery:
      name: telefone
      in: query
      description: telefone.
      schema:
        type: string
      required: false
    emailInQuery:
      name: email
      in: query
      description: email.
      schema:
        type: string
      required: false
    statusInQuery:
      description: 'Filtro de status de usuário'
      in: query
      name: status
      example: 'Refund for fraud'
      required: false
      schema:
        $ref: "#/components/schemas/StatusUsuario"

    authorization:
      name: Authorization
      in: header
      description: A valid OAuth2 access token.
      required: true
      schema:
        type: string
        pattern: Bearer .*
        example: >
          Bearer
          eyJhbGciOiKERzI1NiJ9.eyJleGFtcGxlIjoiZXhhbXBsZSJ9.CIOhRRU_233iQxVsR-4dZH5OGTZgExJAvOx7kUMJlisB